% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRUtils.R
\name{AddMotifs}
\alias{AddMotifs}
\title{Make Signac Motif Object with Granges Object}
\usage{
AddMotifs(object, genome, pfm, verbose = TRUE, ...)
}
\arguments{
\item{object}{Granges object,like peakset from ArchR::getPeakSet}

\item{genome}{BSgenome Object ,eg,BSgenome.Mmusculus.UCSC.mm10}

\item{pfm}{PWMatrixList object,eg,chromVARmotifs::mouse_pwms_v2}
}
\value{
motif.matrix,and you can plot with like Signac::MotifPlot(motif.matrix,motifs =c("Fosb","Mga")
}
\description{
Make Signac Motif Object with Granges Object
}
