% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamToCount.R
\name{bam2Count}
\alias{bam2Count}
\title{bam file to count matrix}
\usage{
bam2Count(
  files = NULL,
  Names = NULL,
  peaks = NULL,
  format = "bam",
  addCol = "Sample",
  paired = TRUE,
  by_rg = TRUE
)
}
\arguments{
\item{files}{bame files after aligment from fastq file}

\item{Names}{relative bam file names}

\item{peaks}{GRanges object}

\item{format}{bed or bam format}

\item{addCol}{colname add into colData}

\item{paired}{whether paried-seq}

\item{by_rg}{use RG tags in bam to separate groups}
}
\value{
‘RangedSummarizedExperiment-class’ object
}
\description{
bam file to count matrix
}
