% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoteBED.R
\name{callSummitsMACS2}
\alias{callSummitsMACS2}
\title{call macs2 for bam file or bed file(pass)}
\usage{
callSummitsMACS2(
  File = NULL,
  format = "BED",
  pathToMacs2 = "macs2",
  genomeSize = 2.7e+09,
  shift = -75,
  extsize = 150,
  cutOff = 0.05,
  method = "q",
  additionalParams = "--nomodel --nolambda",
  outdir = "result",
  prefix = "macs2",
  save = TRUE,
  cleanUp = FALSE
)
}
\description{
call macs2 for bam file or bed file(pass)
}
