% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipSeeker.R
\name{chipseekerGO}
\alias{chipseekerGO}
\title{GO Pathway analysis for peakAnnoList from ChipSeeker Annotation}
\usage{
chipseekerGO(
  peakAnnoList = NULL,
  fun = "enrichKEGG",
  genome = "hg38",
  plot = TRUE,
  outdir = "chipSeeker"
)
}
\arguments{
\item{peakAnnoList}{peakAnnoList object from annotatePeak}

\item{fun}{fun: One of "groupGO", "enrichGO", "enrichKEGG", "enrichDO" or "enrichPathway"}

\item{genome}{genome name,hg38 or mm10}

\item{plot}{whether plot}

\item{outdir}{output path to save result}
}
\value{
compareCluster object
}
\description{
GO Pathway analysis for peakAnnoList from ChipSeeker Annotation
}
