% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2Disease.R
\name{chromVARSNPpipeline}
\alias{chromVARSNPpipeline}
\title{main pipeline(pass)}
\usage{
chromVARSNPpipeline(
  obj = NULL,
  umapMat = NULL,
  SNP = NULL,
  binarize = TRUE,
  groupBy = NULL,
  cutoff = 0.35,
  genome = "hg38",
  outdir = NULL,
  save = FALSE
)
}
\arguments{
\item{obj}{RangedSummarizedExperiment object(eg,from function seTocicero)}

\item{umapMat}{umap Embedding matrix or data.frame}

\item{binarize}{whether binarize data}

\item{cutoff}{float value,filter connection}

\item{genome}{genome name(hg38 or mm10)}

\item{save}{whether save result}

\item{snp}{SNP GRanges object}
}
\description{
main pipeline(pass)
}
