% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{computeBgdPeaks}
\alias{computeBgdPeaks}
\title{chromVar method to caculate Backgroup Peaks}
\usage{
computeBgdPeaks(
  rseObj = NULL,
  Seqnames = NULL,
  nIterations = 50,
  w = 0.1,
  binSize = 50,
  seed = 1,
  outFile = "Background-Peaks.rds"
)
}
\arguments{
\item{rseObj}{RangedSummarizedExperiment Object}

\item{Seqnames}{seqnames use}
}
\value{
RangedSummarizedExperiment Object
}
\description{
chromVar method to caculate Backgroup Peaks
}
\examples{
data(example_counts, package = "chromVAR")
example_counts=addGCBias(example_counts,genome="BSgenome.Hsapiens.UCSC.hg19")
.computeBgdPeaks(example_counts)
}
