% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRUtils.R
\name{computeDORCKNN}
\alias{computeDORCKNN}
\title{compute matrix KNN object}
\usage{
computeDORCKNN(
  data,
  k = 50,
  knnIteration = 500,
  includeSelf = FALSE,
  overlapCutoff = 0.8,
  ...
)
}
\arguments{
\item{data}{matrix data}

\item{k}{integer number for KNN}
}
\description{
compute matrix KNN object
}
