% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateGeneAnnotation.R
\name{createGeneAnnotation}
\alias{createGeneAnnotation}
\title{Create a gene annotation object}
\usage{
createGeneAnnotation(
  genome = NULL,
  TxDb = NULL,
  OrgDb = NULL,
  genes = NULL,
  exons = NULL,
  TSS = NULL,
  annoStyle = NULL
)
}
\arguments{
\item{genome}{genome name(eg,hg38)}

\item{TxDb}{TxDb name}

\item{OrgDb}{OrgDb name
...}
}
\description{
Create a gene annotation object
}
