% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{DeVariability}
\alias{DeVariability}
\title{get RangedSummarizedExperiment Object Deviation Variability}
\usage{
DeVariability(
  seCounts = NULL,
  matchesAnno = NULL,
  genome = "BSgenome.Hsapiens.UCSC.hg19",
  n = 25,
  plot = TRUE
)
}
\arguments{
\item{seCounts}{RangedSummarizedExperiment Object}

\item{matchesAnno}{motif matches object from function `makeMotifAnnotations`}

\item{genome}{genome name,default:BSgenome.Hsapiens.UCSC.hg19}

\item{n}{number of top rank motifs to show}

\item{plot}{whether plot default true}
}
\description{
get RangedSummarizedExperiment Object Deviation Variability
}
\examples{
data(example_counts, package = "chromVAR")
peakset=rowRanges(example_counts)
out=makeMotifAnnotations(peakSet=peakset,genome="BSgenome.Hsapiens.UCSC.hg19")
p=plotDeVariability(seCounts=example_counts,matchesAnno=out[["motifMatches"]])
}
