% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2Disease.R
\name{DevHeatmap}
\alias{DevHeatmap}
\title{plot Deviations}
\usage{
DevHeatmap(dev = NULL, groupBy = NULL, plot = TRUE)
}
\arguments{
\item{dev}{Deviations object}

\item{groupBy}{column in metadata as group,and plot with Diease}

\item{plot}{whether plot}
}
\description{
plot Deviations
}
