% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{.getRowSums}
\alias{.getRowSums}
\title{get row Sums for RangedSummarizedExperiment object}
\usage{
.getRowSums(
  rseObj = NULL,
  Seqnames = NULL,
  verbose = TRUE,
  filter0 = FALSE,
  threads = 1
)
}
\arguments{
\item{rseObj}{RangedSummarizedExperiment object}

\item{seqnames}{seqnames use}
}
\description{
get row Sums for RangedSummarizedExperiment object
}
