% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoteBED.R
\name{extendGR}
\alias{extendGR}
\title{extend Granges object}
\usage{
extendGR(gr = NULL, upstream = 50000, downstream = 50000)
}
\arguments{
\item{gr}{Granges object}

\item{upstream}{upstream value}

\item{downstream}{downstream value}
}
\description{
extend Granges object
}
