% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoteBED.R
\name{fastAnnoPeaks}
\alias{fastAnnoPeaks}
\title{annotate peak Grange Object}
\usage{
fastAnnoPeaks(
  peaks = NULL,
  BSgenome = NULL,
  geneAnnotation = NULL,
  promoterRegion = c(2000, 100)
)
}
\arguments{
\item{peaks}{peak Grange object}

\item{BSgenome}{BSgenome object}

\item{geneAnnotation}{geneAnnotation object}
}
\description{
annotate peak Grange Object
}
