% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{getDeviationsMatrix}
\alias{getDeviationsMatrix}
\title{add Deviation Matrix}
\usage{
getDeviationsMatrix(
  rseObj = NULL,
  matches = NULL,
  prefix = "",
  out = c("deviations", "z"),
  threads = 1,
  verbose = TRUE
)
}
\arguments{
\item{rseObj}{RangedSummarizedExperiment Object}

\item{matches}{object from makeMotifAnnotations motifMatches}

\item{threads}{number of threads}
}
\value{
SummarizedExperiment object
}
\description{
add Deviation Matrix
}
