% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRUtils.R
\name{getDORCXMatrix}
\alias{getDORCXMatrix}
\title{get DORCX matrix from ArchProject with addPeak2Link}
\usage{
getDORCXMatrix(
  project = NULL,
  colCellType = "Clusters",
  dorcCutoff = 10,
  eps = 10000,
  corCutOff = 0.45
)
}
\arguments{
\item{project}{ArchR Project object}

\item{colCellType}{column in metadata as celltype}

\item{dorcCutoff}{the cutoff value for DORC score,default 10}

\item{eps}{value to multiply peakcounts,default 1e4}

\item{corCutOff}{correlation value cutoff for getPeak2GeneLinks,default 0.45}
}
\description{
get DORCX matrix from ArchProject with addPeak2Link
}
\examples{
dorcList=getDORCXMatrix(projHeme)
x=dorcList[["CellTypeDORC"]]
colData=data.frame("Clusters"=colnames(x))
ArchRHeatmap(mat=as.matrix(x),colData=colData,showRowDendrogram=TRUE,scale=TRUE,customRowLabel =c(1,3,5,7))
}
