% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features2Motif.R
\name{GetMotifData}
\alias{GetMotifData}
\title{Get Motif Data}
\usage{
\method{GetMotifData}{Motif}(object, slot = "data", ...)
}
\arguments{
\item{slot}{Information to pull from object (data, pwm, meta.data)}
}
\description{
Get Motif Data
}
\examples{
motif.obj <- GetMotifData(
  object = atac_small[['peaks']], slot = "motifs"
)
GetMotifData(object = motif.obj)
}
\concept{motifs}
