% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features2Motif.R
\name{getMotifsDB}
\alias{getMotifsDB}
\title{get Motifs DataBase}
\usage{
getMotifsDB(species = "Mus musculus", collection = "CORE", ...)
}
\arguments{
\item{species}{which species to use(eg,"Mus musculus" or "Homo sapiens"),default "Mus musculus}

\item{collection}{which collection to use,deafault CORE}
}
\value{
PWMatrixList Object
}
\description{
get Motifs DataBase
}
