% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link2TF.R
\name{getP2GLinks}
\alias{getP2GLinks}
\title{caculate P2G Link ,please reference to (https://github.com/GreenleafLab/MPAL-Single-Cell-2019)(pass)}
\usage{
getP2GLinks(
  seA = NULL,
  seB = NULL,
  corCutOff = 0.2,
  fdrCutOff = 0.1,
  distCutOff = 2500,
  heathP2G = NULL,
  diseaseP2G = NULL,
  totalP2G = NULL,
  GTF = "gene.gtf",
  ...
)
}
\arguments{
\item{seA}{seATAC RangedSummarizedExperiment Object,can get from ArchR::getMatrxFromProject (after use function gr2Feature)}

\item{seB}{seRNA  RangedSummarizedExperiment Object(after use function seRNA2Rse)}

\item{corCutOff}{Pearson Correlation Cutoff,default 0.2}

\item{fdrCutOff}{FDR Cutoff,default 0.1}

\item{distCutOff}{Min Dist to TSS,default 2500}

\item{heathP2G}{can get from `metadata(ArchRProj@peakSet)$Peak2GeneLinks`}

\item{diseaseP2G}{same as heathP2G}

\item{totalP2G}{same as heathP2G}

\item{gtf}{gene gtf file or gene Granges object}
}
\description{
caculate P2G Link ,please reference to (https://github.com/GreenleafLab/MPAL-Single-Cell-2019)(pass)
}
