% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2Disease.R
\name{getTxDbGenes}
\alias{getTxDbGenes}
\title{get TxDb genes}
\usage{
getTxDbGenes(txdb = NULL, orgdb = NULL, gr = NULL, ignore.strand = TRUE)
}
\arguments{
\item{txdb}{txdb oject}
}
\description{
get TxDb genes
}
