% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{gr2Feature}
\alias{gr2Feature}
\title{rename RangedSummarizedExperiment seATAC}
\usage{
gr2Feature(seATAC = NULL)
}
\arguments{
\item{seATAC}{RangedSummarizedExperiment Object(eg,from getMatrixFromProject )}
}
\description{
rename RangedSummarizedExperiment seATAC
}
