% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2Disease.R
\name{gwasSNPconn}
\alias{gwasSNPconn}
\title{connect SNP to scATAC Object}
\usage{
gwasSNPconn(
  se = NULL,
  snp = NULL,
  conn = NULL,
  cutoff = 0.35,
  genome = "hg38",
  outdir = NULL,
  save = TRUE
)
}
\arguments{
\item{se}{RangedSummarizedExperiment object(eg,from function seTocicero)}

\item{snp}{SNP GRanges object}

\item{conn}{connections object from cicero computation}

\item{cutoff}{float value,filter connection}

\item{save}{whether save result}
}
\description{
connect SNP to scATAC Object
}
