% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoteBED.R
\name{identifyReproduciblePeaks}
\alias{identifyReproduciblePeaks}
\title{identify Reproducible Peaks(pass)}
\usage{
identifyReproduciblePeaks(
  summitFiles = NULL,
  summitNames = NULL,
  reproducibility = 0.51,
  extendSummits = 250,
  blacklist = NULL,
  prefix = NULL
)
}
\arguments{
\item{summitFiles}{summitFiles after macs2 callpeak}

\item{summitNames}{relative names for summitFiles}

\item{reproducibility}{reproducibility value}

\item{extendSummits}{extendSummits number}

\item{blacklist}{blacklist Granges object}
}
\description{
identify Reproducible Peaks(pass)
}
