% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{keepFilteredChromosomes}
\alias{keepFilteredChromosomes}
\title{filter chroms}
\usage{
keepFilteredChromosomes(
  x,
  remove = c("chrM"),
  underscore = TRUE,
  standard = TRUE,
  pruning.mode = "coarse"
)
}
\description{
filter chroms
}
