% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link2TF.R
\name{link2TF}
\alias{link2TF}
\title{link peak to TF}
\usage{
link2TF(
  diffRNA = NULL,
  diffATAC = NULL,
  p2gLinks = NULL,
  matches = NULL,
  log2fc = 0.1,
  pval = 0.05
)
}
\arguments{
\item{diffRNA}{object from ArchR::getMarkerFeatures(ArchRProject,useMatrix="GeneScoreMatrix" or "GeneIntegraMatrix")}

\item{diffATAC}{object from ArchR::getMarkerFeatures(ArchRProject,useMatrix="PeakMatrix")}

\item{p2gLinks}{object from `getP2GLinks` function}

\item{matches}{object from ArchR::getMatches}
}
\description{
link peak to TF
}
