% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2Disease.R
\name{makeCDS}
\alias{makeCDS}
\title{RangedSummarizedExperiment objects to newCellDataSet}
\usage{
makeCDS(se, binarize = TRUE)
}
\arguments{
\item{se}{RangedSummarizedExperiment Object}

\item{binarize}{whether binarize matrix}
}
\description{
RangedSummarizedExperiment objects to newCellDataSet
}
