% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{PeakSet2BW}
\alias{PeakSet2BW}
\title{export GRanges peakset object into BW file}
\usage{
PeakSet2BW(gr = NULL, outdir = "bwFile")
}
\arguments{
\item{gr}{GRanges Object}
}
\description{
export GRanges peakset object into BW file
}
