% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{seRNA2Rse}
\alias{seRNA2Rse}
\title{Convert scRNA SummarizedExperiment Object into RangedSummarizedExperiment Object}
\usage{
seRNA2Rse(seRNA = NULL, gtf = "gene.gtf", tssWindow = 2500)
}
\arguments{
\item{seRNA}{SummarizedExperiment Object}

\item{tssWindow}{tssWindow's number,defaut 2500}

\item{gtfFile}{gene gtf file}
}
\value{
RangedSummarizedExperiment object
}
\description{
Convert scRNA SummarizedExperiment Object into RangedSummarizedExperiment Object
}
