% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2Disease.R
\name{seTocicero}
\alias{seTocicero}
\title{SummarizedExperiment to Cicero object}
\usage{
seTocicero(
  obj = NULL,
  umapMat = NULL,
  flank = 250 * 10^3,
  binarize = TRUE,
  save = FALSE,
  outdir = NULL
)
}
\arguments{
\item{obj}{RangedSummarizedExperiment Object(eg.getMatrixFrom(ArchProject,useMatrix="PeakMatrix"))}

\item{umapMat}{umap Embedding matrix or data.frame}

\item{binarize}{whether binarize data}

\item{genome}{genome name(hg38 or mm10)}
}
\description{
SummarizedExperiment to Cicero object
}
