% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{seuratTose}
\alias{seuratTose}
\title{Convert Seurat object into SummarizedExperiment Object}
\usage{
seuratTose(object = NULL)
}
\arguments{
\item{object}{seurat object}
}
\value{
SummarizedExperiment Object
}
\description{
Convert Seurat object into SummarizedExperiment Object
}
