% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{validBSgenome}
\alias{validBSgenome}
\title{get valid BSgenome}
\usage{
validBSgenome(genome = NULL, masked = FALSE)
}
\arguments{
\item{genome}{genome name,eg.BSgenome.Hsapiens.UCSC.hg19
export}
}
\description{
get valid BSgenome
}
