% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoteBED.R
\name{annoteFile}
\alias{annoteFile}
\title{annotate bed files}
\usage{
annoteFile(
  Files = NULL,
  Names = NULL,
  blacklist = NULL,
  format = "BED",
  genome = "hg38",
  outdir = "annotation"
)
}
\arguments{
\item{Files}{vector of bedfile}

\item{Names}{relative bedfile names}

\item{blacklist}{blacklist GRanges object}

\item{format}{input data format}

\item{genome}{genome name}

\item{outdir}{outdir to save result}
}
\description{
annotate bed files
}
