% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamToCount.R
\name{bam2Count}
\alias{bam2Count}
\title{bam file to count matrix}
\usage{
bam2Count(
  bamfiles = NULL,
  bamNames = NULL,
  peaks = NULL,
  addCol = "Sample",
  paired = TRUE
)
}
\arguments{
\item{bamfiles}{bame files after aligment from fastq file}

\item{bamNames}{relative bam file names}

\item{peaks}{GRanges object}

\item{addCol}{colname add into colData}

\item{paired}{whether paried-seq}
}
\value{
‘RangedSummarizedExperiment-class’ object
}
\description{
bam file to count matrix
}
